:recycle: (interaction handler, client utils) clean index, move utils
