+++ TODOBOT/modules/util/emojis.js
@@ -1,0 +1,40 @@
+module.exports = (client) => {
+    client.emojiMap = {
+        1: '1️⃣',
+        2: '2️⃣',
+        3: '3️⃣',
+        4: '4️⃣',
+        5: '5️⃣',
+        6: '6️⃣',
+        7: '7️⃣',
+        8: '8️⃣',
+        9: '9️⃣',
+        10: '🔟',
+        '-': '🟥',
+        '+': '🟩',
+        'edit': '<:edit:820577055342985216>',
+        'finish': '<:finish:820576533059600394>',
+        'assign': '<:assign_yourself:820577858081521675>',
+        'github': '<:github:820639615990890496>',
+        'share': '<:share:820419979719344139>',
+        'task_open': '<:task_open:820381667881517118>',
+        'task_finished': '<:task_finished:820384679562838046>',
+        'upvote': '<:upvote:820678243828105216>',
+        'downvote': '<:downvote:820677972645642290>',
+        'expand': '<:expand:822466806211543080>',
+        'collapse': '<:collapse:822467028983873606>',
+        'accept': '<:accept_todo:822495794602442814>'
+    }
+    client.Mapemoji = {
+        '1️⃣': 1,
+        '2️⃣': 2,
+        '3️⃣': 3,
+        '4️⃣': 4,
+        '5️⃣': 5,
+        '6️⃣': 6,
+        '7️⃣': 7,
+        '8️⃣': 8,
+        '9️⃣': 9,
+        '🔟': 10,
+    }
+}
