+++ TODOBOT/modules/util/permissions.js
@@ -1,0 +1,7 @@
+module.exports = (client) => {
+    client.permMap = {
+        'USER': 0,
+        'BOT_USER': 1,
+        'STAFF': 2
+    }
+}
